Fix additional compilation issues from second PhD review
- HighlightRenderer.cs: Fix PathResult.Success -> PathResult.Reachable
- UnitRenderer.cs: Fix IPoolable method names (OnAcquire/OnRelease ->
  OnGetFromPool/OnReturnToPool)
- SelectionManager.cs: Cast float _movementCost to int for MoveUnit call
- HexGrid.cs: Add IService implementation for ServiceLocator compatibility
- SaveManager.cs: Fix CurrentState.Name -> CurrentState.ToString()
- MapGenerator.cs: Add missing System.Threading using statement

+++ godot/src/CSharp/Input/SelectionManager.cs
@@ -393,7 +393,7 @@ public class PathMoveCommand : ICommand
     public bool Execute()
     {
         var destination = _path[^1];
-        return _unitManager.MoveUnit(_unit.Id, destination.Q, destination.R, _movementCost);
+        return _unitManager.MoveUnit(_unit.Id, destination.Q, destination.R, (int)_movementCost);
     }
 
     public bool Undo()
